[#3] feat: 책 기본 API 작업 완료

+++ src/main/java/com/example/allthatbooks/domain/book/service/BookService.java
@@ -1,8 +1,10 @@
 package com.example.allthatbooks.domain.book.service;
 
+import com.example.allthatbooks.common.dto.response.PageResponse;
 import com.example.allthatbooks.domain.book.dto.request.CreateBookRequest;
 import com.example.allthatbooks.domain.book.dto.request.UpdateBookRequest;
 import com.example.allthatbooks.domain.book.dto.response.BookDetailResponse;
+import com.example.allthatbooks.domain.book.dto.response.BookListResponse;
 import com.example.allthatbooks.domain.book.dto.response.BookSingleResponse;
 import com.example.allthatbooks.domain.book.dto.response.BookTagResponse;
 import com.example.allthatbooks.domain.book.entity.Book;
@@ -11,14 +13,21 @@ import com.example.allthatbooks.domain.book.entity.BookTag;
 import com.example.allthatbooks.domain.book.repository.BookDetailRepository;
 import com.example.allthatbooks.domain.book.repository.BookRepository;
 import com.example.allthatbooks.domain.book.repository.BookTagRepository;
-import com.example.allthatbooks.domain.common.dto.ImageUrl;
-import com.example.allthatbooks.domain.common.enums.Tag;
+import com.example.allthatbooks.common.dto.ImageUrl;
+import com.example.allthatbooks.common.enums.Tag;
+import com.example.allthatbooks.common.exception.CustomException;
+import com.example.allthatbooks.common.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
+@Slf4j
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
 @Service
@@ -43,31 +52,47 @@ public class BookService {
         }
 
         Book savedBook = bookRepository.save(book);
-        List<BookTagResponse> tagResponse = savedBook.getBookTagList()
-            .stream()
-            .map(BookTagResponse::toDto)
-            .toList();
-        List<BookDetailResponse> detailResponse = savedBook.getBookDetailList()
-            .stream()
-            .map(BookDetailResponse::toDto)
-            .toList();
+        return BookSingleResponse.toDto(savedBook);
+    }
 
-        return BookSingleResponse.toDto(savedBook, tagResponse, detailResponse);
+    public PageResponse<BookListResponse> findAll(int page, int size) {
+        Pageable pageable = PageRequest.of(Math.max(0, page - 1), size);
+        Page<BookListResponse> results = bookRepository.findAllPaging(pageable);
+        return PageResponse.toDto(results);
+    }
+
+    public BookSingleResponse findOne(Long bookId) {
+        Book foundBook = bookRepository.findBookByIdFetchJoinOrElseThrow(bookId);
+
+        if (foundBook.getDeletedAt() != null) {
+            throw new CustomException(ErrorCode.ALREADY_DELETED_BOOK);
+        }
+
+        return BookSingleResponse.toDto(foundBook);
     }
 
     @Transactional
     public void updateBook(Long bookId, UpdateBookRequest request) {
-        Book foundBook = bookRepository.findBookByIdOrElseThrow(bookId);
-
         //TODO: userId 검증 로직 필요
-        foundBook.updateBook(request);
 
+        // fetch Join 적용
+        Book foundBook = bookRepository.findBookByIdFetchJoinOrElseThrow(bookId);
+
+        if (foundBook.getDeletedAt() != null) {
+            throw new CustomException(ErrorCode.ALREADY_DELETED_BOOK);
+        }
+
+        foundBook.updateBook(request);
     }
 
     @Transactional
     public void deleteBook(Long bookId) {
         Book foundBook = bookRepository.findBookByIdOrElseThrow(bookId);
-        foundBook.deleteBook(); // 일단 soft
+        if (foundBook.getDeletedAt() != null) {
+            throw new CustomException(ErrorCode.ALREADY_DELETED_BOOK);
+        }
+
+        foundBook.deleteBook();
     }
 
 }
